requirements.txt
PyYAML~=6.0.1
pymemcache~=4.0.0